# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "getProductsList",
                "description": "Get all Products",
                "operationId": "b9260c010a9221aae5692e152e23556d",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Products"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "createProducts",
                "description": "Create Products",
                "operationId": "aeeea8bec7a4f91d3d640c95917c3743",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Products"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Products"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "getProductsItem",
                "description": "Get Products",
                "operationId": "d13ed5a0e96281d0b6dfb9a1f1d71e56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Products",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Products"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "updateProducts",
                "description": "Update Products",
                "operationId": "151ecd290d767189381da828757b8ccc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Products",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Products"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Products"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "deleteProducts",
                "description": "Delete Products",
                "operationId": "f6dbe64bfe381b447fc99b6307aa0b27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Products",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Products": {
                "required": [
                    "name",
                    "price",
                    "stock"
                ],
                "properties": {
                    "name": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "description": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "price": {
                        "description": "",
                        "type": "number",
                        "format": "number",
                        "readOnly": false,
                        "nullable": false
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}